Use inversify to binding modules
