+++ plugins/ssh-plugin/package.json
@@ -29,7 +29,8 @@
   "dependencies": {
     "axios": "0.21.1",
     "fs-extra": "7.0.1",
-    "@eclipse-che/plugin": "0.0.1"
+    "@eclipse-che/plugin": "0.0.1",
+    "inversify": "^5.0.1"
   },
   "devDependencies": {
     "@eclipse-che/api": "latest",
