+++ plugins/ssh-plugin/src/messages.ts
@@ -1,0 +1,25 @@
+/**********************************************************************
+ * Copyright (c) 2019-2021 Red Hat, Inc.
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ ***********************************************************************/
+
+export const MESSAGE_NEED_RESTART_WORKSPACE =
+  'Che Git plugin can leverage the generated keys now. To make them available in all workspace containers please restart your workspace.';
+
+export const MESSAGE_ENTER_KEY_NAME_OR_LEAVE_EMPTY =
+  'Please provide a hostname (e.g. github.com) or leave empty to setup default name';
+
+export const MESSAGE_CANNOT_GENETARE_SSH_KEY = 'Unable to generate SSH key.';
+
+export const MESSAGE_NO_SSH_KEYS = 'There are no SSH keys.';
+
+export const MESSAGE_GET_KEYS_FAILED = 'Failure to fetch SSH keys.';
+
+export const MESSAGE_PERMISSION_DENIED_PUBLICKEY = 'Failure to clone git project. Permission denied (publickey).';
+
+export const BTN_CONTINUE = 'Continue';
